pyproject.toml
[project]
name = "llm-judge"
version = "0.1.0"
description = "LLM asymmetry test suite that scores model responses via OpenRouter."
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
    "openai>=1.13.3",
    "httpx[http2]>=0.25.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4",
]
dev = [
    "black>=23.12",
    "flake8>=6.1",
    "mypy>=1.8",
    "pyright>=1.1",
    "pytest>=8.4.2",
    "types-colorama",
    "types-PyYAML",
]

[build-system]
requires = ["uv-build>=0.3.0"]
build-backend = "uv_build"

[tool.uv.sources]
llm_judge = { path = "src/llm_judge" }

[tool.black]
line-length = 120
target-version = ["py310"]
preview = true

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
strict_optional = true
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
no_implicit_optional = true
strict_equality = true

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.10"
reportMissingImports = "error"
reportMissingTypeStubs = "error"
reportUnknownParameterType = "warning"
reportUnknownArgumentType = "warning"
reportUnknownVariableType = "warning"
reportUnknownMemberType = "warning"
reportUntypedFunctionDecorator = "warning"
reportUntypedClassDecorator = "warning"
reportUntypedNamedTuple = "warning"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-config",
    "--strict-markers",
]
